requirements.txt
python-dotenv
requests
docopts
